Update Dotfuscator renaming exclusions
Most identifiers can be renamed now that Dotfuscator has been upgraded
to version 4.33, which handles Xamarin code constructs better.
There are two exceptions:

1) We don't remove the renaming exclusions from the UWP config file,
because of a known issue where Dotfuscator cannot determine if
certain XAML elements are Xamarin controls or UWP controls.

2) We also continue to exclude BugSweeper.Board.FlaggedTileCount from
renaming. This is because the value of this property changes
during runtime, calling OnPropertyChanged("FlaggedTileCount") to
tell the UI to update. Dotfuscator does not currently modify that
string constant when renaming the property.