Andorid: Enhance protection
Disables library mode on the libraries BugSweeper.dll and
BugSweeperTile.dll so more gets renamed.

Configures a renaning exclusion on the type BugSweeper.BugSweeperPage
so the app still runs when the libraries are no longer in library mode.

Enables Enhanced Overload Induction to further improve renaming
protection

Enables Supressildasm option to prevent decompilation

Enables String Encryption on all the assemblies in the project