Android: Integrate Dotfuscator into project
Imports the PreEmptive.Dotfuscator.Xamarin.targets file on the
last line before the close of the </project> tag.

Configures the following settings in a new property group:
- Sets the DotfuscatorXamarinConfigFileName to 'DotfuscatorConfig.xml'
- Sets the DotfuscatorXamarinCliPath to the location of the Dotfuscator
Community CLI.
- Sets 'DotfuscatorXamarinGenerateNewConfigFile' to 'true' so a
Dotfuscator config file will be generated for the project
- Configures Dotfuscator to protect builds in the Release
configuration by setting 'DotfuscatorXamarinEnabled' to 'true' for
that configuration

Adds the DotfuscatorConfig.xml File to the project so it will be tracked
in Visual Studio.