Andorid: Add renaming exclusions
Exclusions are configured by opening DotfuscatorConfig.xml in
either Dotfuscator Community or the Dotfuscator Professional Config
Editor.

BugSweeperPage and the board field were excluded so that the app will
run again

The FlaggedTileCount is excluded so the value goes back to being updated
during the game.

The BugCount is excluded so it goes back to being displayed.