Andorid: Configure Root Check
Puts the root check on the "OnSingleTap" method and sets the check
action to "Exit" so that the app will exit when that method is called
on a rooted device.

To test the root check, open the app on a rooted emulator or
rooted device and try flagging one of the tiles as bug. The app should
exit.