Android: Add tamper check
Puts the tamper check on the "NewGameInitialize" method and set the check
action to "Exit" so othat the app will exit when that method is called
in a tampered apk.

Enables Xamarin app signing and sets "DotfuscatorAndroidSigningCertFingerprint"
property in Bugsweeper.Android.csproj. The fingerprint can be discovered
from an existing keystore by using "keytool.exe -list". Keytool can be found
inside the Java Development Kit installation folder.

To test the tamper check, either resign the apk with a different certificate
or disable signing. This simulates an attacker making changes to the apk
and resigning those changes. Open the app, and it should exit immediately.